Amend and add to documentation on DSL Joins
1. Remove invalid example.

```kotlin
(Players innerJoin StarWarsFilms)
  .slice(Players.name.count(), StarWarsFilms.name)
  .select { StarWarsFilms.sequelId eq Players.sequelId }
  .groupBy(StarWarsFilms.name)
```

The `innerJoin` line will throw an `IllegalStateException` with the message `Cannot join with StarWarsFilms as there is no matching primary key/foreign key pair and constraint missing`.

2. Add `Roles` table to illustrate foreign key joining.

Example uses `innerJoin` infix and shows the equivalent full form.

3. Rename `Players` to `Actors` throughout.

"Players" is not incorrect, but "Actors" is more common in most varieties of English when referring to movie actors, and this matches the DAO page.

4. Grammar/clarity edits.